+++ src/token.c
@@ -4,11 +4,11 @@
 #include "token.h"
 
 
-Token_Type resolve_token(char* buf, size_t size) {
-	size_t len = strnlen(buf, size);
+Token_Type resolve_token(char* buf, u32 size) {
+	u32 len = strnlen(buf, size);
 	for (int i = 0; i < __TOK_KEYWORD_END; ++i) {
 		const char* tok = token_value[i];
-		size_t token_len = strlen(tok);
+		u32 token_len = strlen(tok);
 		if (strncmp(tok, buf, token_len) == 0 && len == token_len) {
 			return i;
 		}
